refactor: restructure to parallelize document write
